+++ packages/rujira/src/client.ts
@@ -10,7 +10,7 @@ import { RujiraAssets } from './modules/assets.js'
 import { RujiraDeposit } from './modules/deposit.js'
 import { RujiraGhost } from './modules/ghost.js'
 import { RujiraOrderbook } from './modules/orderbook.js'
-import { RujiraPerps } from './modules/perps.js'
+import { RujiraRange } from './modules/range.js'
 import { RujiraStaking } from './modules/staking.js'
 import { RujiraSwap, type RujiraSwapOptions } from './modules/swap.js'
 import { RujiraWithdraw } from './modules/withdraw.js'
@@ -41,7 +41,7 @@ export class RujiraClient {
   public readonly withdraw: RujiraWithdraw
   public readonly staking: RujiraStaking
   public readonly ghost: RujiraGhost
-  public readonly perps: RujiraPerps
+  public readonly range: RujiraRange
   public readonly discovery: RujiraDiscovery
 
   private queryClient: CosmWasmClient | null = null
@@ -79,7 +79,7 @@ export class RujiraClient {
     this.withdraw = new RujiraWithdraw(this)
     this.staking = new RujiraStaking(this)
     this.ghost = new RujiraGhost(this)
-    this.perps = new RujiraPerps(this)
+    this.range = new RujiraRange(this)
 
     this.discovery = new RujiraDiscovery({
       rpcEndpoint: this.config.rpcEndpoint,