+++ examples/browser/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @vultisig/example-browser
 
+## 0.1.28
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vultisig/sdk@0.17.1
+  - @vultisig/examples-shared@0.1.12
+
+## 0.1.27
+
+### Patch Changes
+
+- Updated dependencies [[`219cb00`](https://github.com/vultisig/vultisig-sdk/commit/219cb00898deeaac418945a89c1d243f25aae152)]:
+  - @vultisig/sdk@0.17.0
+  - @vultisig/examples-shared@0.1.11
+
 ## 0.1.26
 
 ### Patch Changes